# specs/1-todo-backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  description: REST API for Todo application with user authentication and task management
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://your-production-domain.com/api
    description: Production server

paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  description: User's password
                  example: securePassword123
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      userId:
                        type: string
                        format: uuid
                        description: The ID of the created user
                        example: "550e8400-e29b-41d4-a716-446655440000"
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Email already exists"

  /auth/login:
    post:
      summary: Authenticate user
      description: Authenticates user with email and password, returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: user@example.com
                password:
                  type: string
                  description: User's password
                  example: securePassword123
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    required:
                      - token
                      - user
                    properties:
                      token:
                        type: string
                        description: JWT token for authentication
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                            description: User ID
                            example: "550e8400-e29b-41d4-a716-446655440000"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Invalid email or password format"
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Invalid credentials"

  /tasks:
    get:
      summary: Get user's tasks
      description: Retrieves all tasks for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      tasks:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Unauthorized"
    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  description: Task title
                  example: "Buy groceries"
                description:
                  type: string
                  maxLength: 1000
                  description: Optional task description
                  example: "Milk, eggs, bread"
                status:
                  type: string
                  enum: [pending, completed]
                  default: pending
                  description: Task status
                  example: "pending"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Title is required"
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Unauthorized"

  /tasks/{id}:
    put:
      summary: Update a task
      description: Updates an existing task for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task ID
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  description: Updated task title
                  example: "Buy groceries"
                description:
                  type: string
                  maxLength: 1000
                  description: Updated task description
                  example: "Milk, eggs, bread, fruits"
                status:
                  type: string
                  enum: [pending, completed]
                  description: Updated task status
                  example: "completed"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Invalid status value"
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Unauthorized"
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Task not found"
    delete:
      summary: Delete a task
      description: Deletes an existing task for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task ID
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Task deleted successfully"
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Unauthorized"
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - error
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Task not found"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - status
        - userId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Task unique identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          description: Optional task description
          example: "Milk, eggs, bread"
        status:
          type: string
          enum: [pending, completed]
          description: Task status
          example: "pending"
        userId:
          type: string
          format: uuid
          description: ID of the user who owns the task
          example: "550e8400-e29b-41d4-a716-446655440000"
        createdAt:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2023-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Task last update timestamp
          example: "2023-01-01T00:00:00.000Z"

security:
  - bearerAuth: []

tags:
  - name: Authentication
    description: User registration and login endpoints
  - name: Tasks
    description: Task management endpoints